authentication/serializers.py
from rest_framework_simplejwt.serializers import TokenObtainSerializer
from rest_framework_simplejwt.tokens import RefreshToken


class FingerprintTokenObtainPairSerializer(TokenObtainSerializer):
    token_class = RefreshToken

    def validate(self, attrs):
        data = super().validate(attrs)
        refresh = self.get_token(self.user)
        refresh["fingerprint"] = "hash"
        data["refresh"] = str(refresh)
        data["access"] = str(refresh.access_token)

        return data
