requirements.txt
Django~=5.0.1
python-dotenv~=1.0.0
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.1
PyJWT~=2.8.0
